requirements.txt
blinker==1.9.0
click==8.1.7
dnspython==2.7.0
exceptiongroup==1.2.2
Flask==3.1.0
Flask-JWT-Extended==4.7.1
Flask-PyMongo==2.3.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
PyJWT==2.10.1
pymongo==4.10.1
pytest==8.3.4
pytest-flask==1.3.0
python-dotenv==1.0.1
tomli==2.2.1
Werkzeug==3.1.3
